Fix thermal-engine dac_override denial
We occasionally see a dac_override SELinux denial from thermal-engine
trying to access /dev/diag.  This seems to be a race between
thermal-engine reading the file and init setting its permissions to
666 (they are initially set to 660 and in the diag group).

This commit adds thermal-engine to the diag group so that it can read
the file even if it is not world-readable.